Fix: Missing null check in LRU cleanup (#87124)
Unfortunately this field is not strongly typed because it's a cyclic
type and I wasn't able to soundly model it with TypeScript. Skill issue
I'm sure.

Don't have an exact repro yet but the previous code was obviously wrong.

The way this should have been structured regardless is to delete the
wrapper MapEntry object instead of the value it contains. The
non-nullness of the wrapper object _is_ properly typed, and avoids an
unnecessary indirection.

As a follow up, I'll look into how to model this type properly.